Halfway there to getting a gameBoard, which works exclusively with links, and a viewer that shows wikipedia content that it uninteractable. The reasoning for this is that a simple removal of links that I don't want is much easier to do when you have more robust, less flimsy data coming from wikipedia's API.
So far, we have the viewer working, it is showing content. we also have the links being ingested, but we still need to write the link code, to show it on the page. then, we will need to add a few buttons for the link. one to update the viewer, and one to update the game board.